requirements.txt
zdesk
doit
PyYaml
urllib3